Improve support for detecting the a system mill installation
The current approach to detect a system-wide mill installation and its
version broke when mill changed the universal script to no longer
include the version information. Therefore, millw was unable to
extract the version from the mill binary.

We now switch to a more robust approach to extract the version of the
mill binary, by executing it. Executing a binary is significantly
slower than inspecting the strings within a binary to detect the mill
version. Therefore we cache the found version.